thesis results almost done

+++ Testgen/Testgen_Sonam/Testgen/spectral-norm.c
@@ -10,9 +10,9 @@
 #include <stdlib.h>
 #include <math.h>
 
-double eval_A(int i, int j) { return 1.0/((i+j)*(i+j+1)/2+i+1); }
+float eval_A(int i, int j) { return 1.0/((i+j)*(i+j+1)/2+i+1); }
 
-void eval_A_times_u(int N,  double u[], double Au[])
+void eval_A_times_u(int N,  float u[], float Au[])
 {
   int i,j;
   for(i=0;i<N;i++)
@@ -22,7 +22,7 @@ void eval_A_times_u(int N,  double u[], double Au[])
     }
 }
 
-void eval_At_times_u(int N,  double u[], double Au[])
+void eval_At_times_u(int N,  float u[], float Au[])
 {
   int i,j;
   for(i=0;i<N;i++)
@@ -32,15 +32,15 @@ void eval_At_times_u(int N,  double u[], double Au[])
     }
 }
 
-void eval_AtA_times_u(int N,  double u[], double AtAu[])
-{ double v[N]; eval_A_times_u(N,u,v); eval_At_times_u(N,v,AtAu); }
+void eval_AtA_times_u(int N,  float u[], float AtAu[])
+{ float v[N]; eval_A_times_u(N,u,v); eval_At_times_u(N,v,AtAu); }
 
 int main()
 {
   int i;
   int N ;
   scanf("%d",&N);
-  double u[N],v[N],vBv,vv;
+  float u[N],v[N],vBv,vv;
   for(i=0;i<N;i++) u[i]=1;
   for(i=0;i<10;i++)
     {
